requirements.txt
Django==3.0.8
djangorestframework==3.11.0
gunicorn==20.0.4
dj-database-url==0.5.0
psycopg2-binary==2.8.6
whitenoise==5.2.0
Pillow==7.2.0
django-cors-headers==3.4.0
django-storages==1.10.1
boto3==1.15.10